pyproject.toml
[tool.poetry]
name = "spotify-market-popularity"
version = "0.1.0"
description = ""
authors = ["Matt Bolanos <matthew.a.bolanos@gmail.com>"]
license = "MIT"
readme = "README.md"

[[tool.poetry.packages]]
include = "functions"

[[tool.poetry.packages]]
include = "data"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pandas = "^2.1.1"
numpy = "^1.26.0"
pre-commit = "^3.4.0"
python-dotenv = "^1.0.0"
ipykernel = "^6.25.2"
selenium = "^4.12.0"
black = "^23.9.1"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
